fix unit test issues with SQLite

+++ src/tests/databases/mock_db/table.sql
@@ -1,13 +1,13 @@
 CREATE TABLE `gibberish` (
-    `id`    INTEGER(11) NOT NULL,
+    `id`    INTEGER NOT NULL,
     `grp`   VARCHAR(20) DEFAULT "",
     `value` INT DEFAULT NULL,
     PRIMARY KEY (`id`, `grp`)
 );
 CREATE INDEX `id_idx` ON `gibberish` (`id`);
 CREATE TABLE `meta` (
-    `meta_id`    INTEGER(11) PRIMARY KEY /*!40101 AUTO_INCREMENT */,
-    `species_id` INTEGER(10) /*!40101 UNSIGNED */ DEFAULT '1',
+    `meta_id`    INTEGER PRIMARY KEY /*!40101 AUTO_INCREMENT */,
+    `species_id` INTEGER /*!40101 UNSIGNED */ DEFAULT '1',
     `meta_key`   VARCHAR(40) NOT NULL,
     `meta_value` VARCHAR(255) NOT NULL
 );
